Move helper function into a new options utilities file
