+++ src/app/static/src/app/store/modelOptions/mutations.ts
@@ -3,6 +3,7 @@ import {ModelOptionsState} from "./modelOptions";
 import {Control, SelectControl, MultiSelectControl, Option, DynamicFormData, DynamicFormMeta} from "@reside-ic/vue-next-dynamic-form";
 import {PayloadWithType} from "../../types";
 import {writeOptionsIntoForm} from "../../utils";
+import {checkOptionsValid} from "./optionsUtils";
 import {VersionInfo, Error, ModelOptionsValidate} from "../../generated";
 
 
@@ -81,53 +82,3 @@ export const mutations: MutationTree<ModelOptionsState> = {
         state.version = action.payload;
     }
 };
-
-function checkOptionsValid(formMeta: DynamicFormMeta): boolean {
-    let valid = true;
-    formMeta.controlSections.forEach(section => {
-        section.controlGroups.forEach(group => {
-            group.controls.forEach(control => {
-                // We could go further and check that if it is null or empty that this isn't
-                // a required control but just assuming this for now seems ok
-                if (control.value != null && control.value != "" && hasOptions(control)) {
-                    valid = valid && checkControlOptionValid(control);
-                }
-            })
-        })
-    })
-    return valid;
-}
-
-type ControlWithOptions = SelectControl | MultiSelectControl;
-
-function hasOptions(control: Control): control is ControlWithOptions {
-    return control.type === "select" || control.type === "multiselect";
-}
-
-function checkControlOptionValid(control: ControlWithOptions): boolean {
-    let valid = true;
-
-    const options = getAllOptions(control)
-    const value = control.value;
-    // Check string and array types, otherwise we assume valid
-    if (typeof value === 'string') {
-        valid = options.includes(value);
-    } else if (Array.isArray(value)) {
-        valid = value.every(item => options.includes(item));
-    }
-
-    return valid;
-}
-
-function getAllOptions(control: ControlWithOptions): string[] {
-    const options= control.options.map((option: Option) => getOptions(option));
-    return options.flat();
-}
-
-function getOptions(option: Option): string[] {
-    const options = [option.id];
-    if (option.children !== undefined) {
-        options.concat(...option.children.map((child: Option) => getOptions(child)));
-    }
-    return options;
-}
